[FIX] Update Font (#176)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Обновляет шрифт

## How This Contributes To The Skyrat Roleplay Experience

<!-- Please add a short description of why you think these changes would
benefit the game and the roleplay atmosphere of the server. If you can't
justify it in words, it might not be worth adding. -->

Красивое

## Proof of Testing

<!-- Include any screenshots/videos/debugging steps of the code
functioning successfully, between the </summary> and </details> code
blocks. -->
<!-- To our mappers and spriters: Posting screenshots of content INSIDE
EDITORS (aseprite, PDN, SDMM, ect) is NOT valid proof of testing. Please
make sure that you COMPILE the game and provide PROOF you tested your
edits. -->

<details>
<summary>Screenshots/Videos</summary>
  
До

![Screenshot_14](https://github.com/ss220club/Skyrat-tg/assets/31931237/abaeb508-4720-459c-96aa-2ef878f4ad28)

![Screenshot_15](https://github.com/ss220club/Skyrat-tg/assets/31931237/9dd02c92-2e89-4446-bf4b-8a8321bc9ab7)

![Screenshot_16](https://github.com/ss220club/Skyrat-tg/assets/31931237/72c354e4-7614-42a6-9c13-3c3e3bf163c0)


После

![Screenshot_11](https://github.com/ss220club/Skyrat-tg/assets/31931237/0df2745a-8f00-44bf-9715-9383a13c1897)

![Screenshot_12](https://github.com/ss220club/Skyrat-tg/assets/31931237/f81e3bd0-e21d-4826-a263-ff3025212232)

![Screenshot_13](https://github.com/ss220club/Skyrat-tg/assets/31931237/3a3c25f0-9e0f-4ceb-b4e3-83c59319f412)

</details>

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
fix: made something easier to use
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->